split it out
